Add the Refactoring
